+++ src/app/reducers/current-list.reducer.ts
